Last touch up to solve this merge
Adds some missing files and restores some missing functionality.
This should include all the different datasources.

Next step would be to combine the different sources, for example under
sphere or commit. But how, when and in what manner well leave to our
iterative process.

I think we have different views on what were trying to accomplish so
I will create issues with some questions/suggestions on how to
proceed please move any discussion on future work to theese.

Also please let me know if anything is missing. The merge was rather big
but as far as I can see all functionality should be included.
Anything missing is an oversight and should be fixed asap.

Please dont forget to sync your individual branches with upstream
regularly to increase our chances of smooth sailing.
