+++ datasets/dummy-students.json
@@ -1,0 +1,57 @@
+[{
+    "city":"Kalmar",
+    "lat":"56.665954",
+    "lng":"16.356824",
+    "created":"2015-08-27T12:41:36.251Z",
+    "firstname":"Johan",
+    "lastname":"Leitet",
+    "program":"staff",
+    "services":{
+      "github":"Leitet"
+    },
+    "studentType":"campus",
+    "username":"rk222ev"
+  },
+  {
+    "city":"Kalmar",
+    "lat":"56.665954",
+    "lng":"16.356824",
+    "created":"2015-08-27T12:47:56.042Z",
+    "firstname":"John",
+    "lastname":"Haggerud",
+    "program":"staff",
+    "services":{
+      "github":"thajo"
+    },
+    "studentType":"campus",
+    "username":"sk222sw"
+  },
+  {
+    "city":"Kalmar",
+    "lat":"56.665954",
+    "lng":"16.356824",
+    "created":"2015-08-27T13:33:33.607Z",
+    "firstname":"Mats",
+    "lastname":"Loock",
+    "program":"staff",
+    "services":{
+      "github":"mtslck"
+    },
+    "studentType":"campus",
+    "username":"mw222rs"
+  },
+    {
+    "city":"Falsterbo",
+    "lat":"55.397107",
+    "lng":"12.841880",
+    "created":"2015-08-27T13:33:33.607Z",
+    "firstname":"Johanna",
+    "lastname":"Szepanski",
+    "program":"student",
+    "services":{
+      "github":"js223kz"
+    },
+    "studentType":"distance",
+    "username":"js223kz"
+  }
+]
