+++ src/services/committer-service.js
@@ -1,14 +1,14 @@
-const studentFile = require("../../datasets/students.json");
-const fs = require('fs');
+const students = require("../../datasets/dummy-students");
 
 module.exports = {
+  students() {
+    return new Promise((resolve, reject) => {
+      resolve(students);
+    });
+  },
   find_by_username(username) {
     return new Promise((resolve, reject) => {
-      studentFile.forEach((student)=>{
-        if(student.services.github === username){
-          resolve(student);
-        }
-      });      
+      resolve(students.find(o => o.username === username));
     });
   }
 };
